+++ src/main/java/com/bts/personalbudget/mapper/FixedBillMapper.java
@@ -22,6 +22,7 @@ public interface FixedBillMapper {
     FixedBill toModel(FixedBillEntity entity);
     FixedBillResponse toResponse(FixedBill model);
     List<FixedBill> toModelList(List<FixedBillEntity> fixedBillEntities);
+    List<FixedBillResponse> toResponseList(List<FixedBillEntity> fixedBillEntities);
     @Mapping(target = "flagActive", constant = "true")
     FixedBillEntity toEntity(FixedBill fixedBill);
